libft: 74% complete of part 1

+++ projects/01_libft/06_strlen/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: jose-ero <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/28 18:20:25 by jose-ero          #+#    #+#             */
-/*   Updated: 2023/04/28 18:34:01 by jose-ero         ###   ########.fr       */
+/*   Updated: 2023/05/01 17:09:58 by jose-ero         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
